Implements Key Value Store Setup
This code implements loading the key value store from the Air Quality JSON file obtained from https://data.cdc.gov/api/views/cjae-szjv/rows.json?accessType=DOWNLOAD.

If there is an existing binary file with the data, then that data is loaded instead.

Once necessary changes (to be implemented) are made, the modified data is written to the binary file. In the future, we would like to only write in the changes.

The data is stored in a dictionary, which holds a tag specifying if the block is free and can be overwritten (which is initialized to false). Other items stored in the values itself is a dictionary which includes an id associated with the measurement, type  of measurement, an id associated with a specific state, that state's name, an id associated with a county, that county's name, year of measurement, and if there are any units, then its name and symbol.