Retrieves input and minor changes
The file name better reflects the project.

Also, we now retrieve input until the user inputs exit, quit, or abort. This input and the key value store is passed to a dedicated function which will be filled in to handle insertions, deletions, updates, selects, and searches.

Finally, the name of the storage file containing the key value store itself has been changed to AirQualityDBStore.bin.