Partially implement Insert to Storage File
A naive version of insert is now implement in which any values insert are simply appended to the end of the file. A version that involves replacing deleted values will likely come after delete has been implemented.

First, 1000-size of row is appended to the end of the file, and then the row itself is written in byte format.

For consistency, all columns will be in camel case, which includes the set.

Additionally, generateNewRows is now implemented in which the tags and column pairs are checked if they are written in a valid format (col=tag, col2=tag, etc). Once that's done, the cols are set as the keys for the tags in each row generated. In a more elaborate explanation, the cols and tags are stored in an array, where every two indices marks a new pair. Once this dictionary is returned, a new dictionary is generated using these values as the data, and with our generated key from the last commit as the key.

There was a bug in generateRandomKey which has now been fixed, which is to force all keys to have only lowercase letters for all letters included.

A minor bug in handleInput is that = will throw an error because we set the structure of values to be (col=tag, col=tag,etc). As a result, we remove the = from an acceptable character in a key or value.

Minor bugs involving comparisons of empty strings as I forgot to set key to be one of the values from the regex comparison are now resolved.