Implement Trivial Selects and some minor fixes
The trivial selects are 1. SELECT * or SELECT * FROM KEYS, which writes keys to a file. All this involves is looping through the key value store dictionary in the running program and printing every key to a text file (printing directly to terminal might be a little problematic as there is a lot of data).
2. SELECT * FROM VALUES, which prints just all values to a text file. This involves looping through the key-value store stored in the dictionary and writing all values to a text file.
3. SELECT * FROM ALL which prints all values to a text file. This involves looping through the key-value dictionary and printing all the keys and all their corresponding values.
4. SELECT [key] or SELECT [key] FROM ALL involves printing the value corresponding to a given key to the consolve. All this involves is first inquring if the key is in the dictionary, and then if it is, then printing out its corresponding value.

Also, the code for handling input specifically for inserts/deletes are refactored into separate functions to make them less unwieldy.

There are more TODO comments and questions, as well as a change to keep all lines to less than or equal to 80 characters.