Change Dictionary Structure and Writing Structure
The key value store also includes an internal variable called position which keeps track of items to be deleted from the key value store.

Additionally, the structure of the key value store is altered to include the padding at the beginning of each block, before the value is written in.