Implement Mostly Working Deletes
Note: This only works for Python 3.7 now as dictionaries are ordered for Python 3.7+.

The structure of deletes involves marking items in the running key value stores as "free," or invalid chunks of data. Additionally, a running list of keys to delete is kept and used to extract indices or rows from the file where we delete the specific key-value pair. The list of such indices is derived and used to delete from the file if the user specifies that they want to quit or exit.

Additionally, handleInput returns both a list of rows to insert and a list of keys to delete.

A new function, updateFileWithDeletes, reads all the lines in the storage file as a list, omits the rows corresponding to the indices of the keys to be deleted, and writes the modifed list back to the storage file. As write overwrites any data to come after the line to be written to, this seemed like the ideal option.

Documentation for all the functions is now for the most part updated.