Add indication of new values
If a user inserted and deleted the same values, then we need to know not to delete from the storage file values that didn't exist yet. And we need to know not to insert to the storage file values that we decided not to insert.

Therefore, the tag position is used for this purpose. We add position tags to indicate when the values were added or updated, and when flagging for insertion and deletion, we check if there is a position tag to indicate if the item is in the storage file and take appropriate action.