Hide metadata
Per suggestion, when displaying any information about the key-value store running in the program, any meta data (including whether the data is invalid or its position tag) is hidden. This is done simply through key referencing, as I, not the user, gets to dictate the names of keys for the metadata.