Modify insertion/deletion/selection to handle local insertions/deletions
isFree is used to signal any local deletions, so in the following cases:

1) If an item was inserted and then deleted, then this item is removed from the rows to be inserted to the storage file.

2) If an item was deleted and a new item was inserted with the same key, then the item will be inserted with that key (as that is basically updating the information and recycling the key).

3) If an item is inserted and deleted in the same cycle, none of its information will be displayed in the existing selects.

All of these are made possible with a simple check in the key value store (in the code, its designated as dynamicDB):
dynamicDB[key]['isFree'] == false.

Additionally, deletion from the dictionary storing insertedRows is done right after the user calls to delete that specific row.