Implementing Updates
Updates of the form UPDATE [key] WITH VALUES (col=tag, col2=tag2, etc) are now parsable and can be used to change the key value store running in the program and in the storage file.

In handleUpdates, regex is used to check the format actually matches the one specified, checking that all tags and columns are matched correctly, and that the column we want to modify is already in the row or in the typesSet. If at least one of these criteria is not met, then we abandon any modification.

handleInput now returns three items, two dictionaries corresponding to inserted or updated rows, and a list of keys corresponding to rows to delete.

In the main function, the updates rows are added to the list of updated rows we will write to the file if the row has not been inserted since the storage file was loaded upon beginning the program. Rows are also removed from the list of updated rows if they are to be deleted.

A new function, updateFileWithUpdates modifies the
key value store from the storage file. As Python 3.7 preserves order in dictionaries, it's possible to recall the positions of keys (in the storage file) to be updated by using simple Python function calls (handled in the main function). The list of tuples with these indices, the key, and its value, is an argument along with the storage file. As the key did not change, the positions that changed are passed in as a list with its associated key and row and then using read lines, we modifiy only the lines we want and then write them back to the file.