Implement a more complex select
This new SELECT statement is of the form SELECT WHERE col=tag, col2=tag2, etc, and regex is used to retrieve its matches. The list of keys is then extracted from a function that is also called by the handleDeletes function, currently named performTempDeletion, which selects for cols and their matching values corresponding to those existing in the dictionary (the logic should be explained in that function and an earlier commit message).

performTempDeletion also has a few changes for potential bugs. One of these are that tags/column descriptions must always be in pairs, otherwise, there is a potential indexError raised. There is also a friendly print message in case there are no rows matching the cols/tags, which reminds the user of valid column names (measureId, measureDesc, stateId, stateName, countyId, countyName, year, measurement, units, unitSymbol).