Implement search, saves/undos, and more
There are 3 major changes and 5 minor changes:

Major Changes:

1. Search is now implemented. The following options are handled:
1. VALUES (col=pattern1, col2=pattern2, col3=pattern3….),
2. VALUES (pattern1, pattern2, pattern3),
3. KEY AND VALUES (pattern1, pattern2, pattern3),
and 4. KEY (pattern1, pattern2,…).
There are two functions handling search: printSearchResult() and handleSearches().
In summary, printSearchResult() this function searches the entire key value store for rows that match the pattern based on the specifications the user provided. handleSearches() checks that the user input is in the form(s) described above and passes necessary values to printSearchResult() and if the forms are not matched, then a message is printed to the user indicating that the input is incorrect and a usage is printed with those following forms.
In more detail, handleSearches(): The above forms are matched on using regex, and based on the key words specified in all caps,  the following arguments are passed to printSearchResult(): two booleans indicating separately if the keys and values, respectively, will be search for necessary patterns, and col names (if specified by the user), and patterns (which as of now are supposed to be substrings of their applicable values). There are a few possible options:
For every pattern, one of the following is checked for every row in the filtered key-value store:
1. If the user specified column names to filter by, then for the pattern’s corresponding column, its corresponding value is checked if any of its substrings are equal to the corresponding pattern provided by the user.
2. Otherwise, the key (if specified) and/or values (if specified) are all scanned if any of their substrings match the pattern per row.
If neither meets the condition, then the row is filtered out. In the end, the results are printed out.
2. Users now have the option to save their changes to the storage file or to abandon, or “undo” them, by specifying “save” or “undo.”
Because there are two spots in the main function which save the file, a newly created function, saveChanges(), saves the changes made to the running key-value store. saveChanges saves the currently running key value store to the storage file (which will be created if necessary). The indices of the deleted keys are stored in a list, which are then passed to updateFileWithDeletes to delete from the storage file. Next, for every "key" that was updated, the index of the key (from the list of keys obtained from the dictionary), the key itself, and its value are stored in a list, which is used to update the corresponding line in the storage file, and finally, the list of inserted rows is passed to the updateFileWithInserts, where the information is written to the end of the file.
handleUpdates returns the new updated row in the key-value store running in the program and the old row (which now makes the third output of handleInput a tuple and not a dictionary). The old row is added to a dictionary keeping track of the data that was replaced in the main function, which only is used when rewriting the old data to the key value store in the case the user specifies that all changes be undone. This would include reverting updates to the old data.
Saving will call saveChanges, and then will remove the items in the running key-value store which are deleted (as the only change made was to mark them as free). Undoing will remove the rows in the running key value store which were inserted, change the ‘isFree’ tag of deleted rows to ‘false’ (to mark them as valid again), and change the ‘data’ value of updated rows to that of the replaced value (which we now keep track of thanks to handleUpdates).
3.
A new function, printSelects(), handles the printing of selects to a file or output.
 First, the user is prompted to indicate whether to print the output to terminal. Next, if the output is large, and the user indicates that they would like the output to be printed to the terminal, the user is notified, and is prompted whether they want to proceed with printing to terminal. In the case that they want to print to file, the user is prompted to indicate if they would like to specify a file to store the output. If not, the given default filename is used. If so, then the program attempts to open or create the file, and if an error is thrown in case the file cannot be created, then, the user is prompted to indicate if they would like to specify a file to store the output. This repeats until the user types in N, or if they specify a valid file path. In any case the user is prompted to type Y or N and they do not, they are prompted again to type Y or N.
As a result, handleSelects no longer writes to files or prints. It simply collects the information that will be printed or written to a file, as well as a default file name depending on the option (allKeys.txt for SELECT * or SELECT * FROM KEYS, allValues.txt for SELECT * FROM VALUES, allKeysValues.txt for SELECT * FROM ALL, and matches.txt for nearly all the other accepted options).

Minor Changes:
1. There is a TODO to read only necessary lines when updating, which will be written back to the file.
2. There is a TODO to read only certain lines when deleting (either the ones to be deleted or the ones that will be written towards the top of the file as records above them will be deleted).
3. Because searches require that all patterns/column names stay together until handleSearches() (because it would be too complicated to check if the user specified VALUES (col=pattern1, col2=pattern2, col3=pattern3….)), handleInput() now includes “=” as a character to be allowed in a regex group, but to avoid affecting the other possible commands, the match groups are generated in the functions which handle the other commands (handleUpdates, handleDeletes, handleInserts, and handleSelects).
4. Because of its unwieldiness, the main function is being refactored into multiple functions. Now, loading the existing storage file or loading the JSON file with the data (if the storage file does not exist yet) are handled in a newly created function, loadFile().
5. In some cases, with invalid input, the individual handleUpdates, handleDeletes, and handleInserts return None objects, so whether each output is None is checked before trying to access output data.