Fix bug for update
In the case that there are multiple updates after a save or loading the program, calls delete and then undo, the original value now displays as it should.

Earlier, because we were trying to store a copy of a dictionary object, any modifications to the object changed the copy itself. Now, at the beginning of handleUpdates(), the object to be replaced is converted to a string and then converted back.

In the main function, when undo is called, we now restore the values from the replacedRows list. replacedRows also does not change if update has been called multiple times.