Update Documentation and Bug Fix
The biggest fix is updating deleting lines from files. While the first line to be deleted was deleted, the other lines were incorrectly deleted. The indices are now updated correctly, as the index is simply the difference between the index of the line to delete (in the file) and the first index deleted.

A minor fix is changing the comment(s) that this program is supported by versions 3.7+ to 3.6+, as order preservation first appears in Python 3.6+. The program will only run if the installation version is at least 3.7.

Additionally, parameter and return values are further identified and described in the docstring.

Some print statements are added as well, to provide further information, including that x items were found, during and after insertion/deletion/updating/searching/selecting.

Additionally, the print statements are updated for consistency, so now tags are column types.

Furthermore, we generalize the key value store to no longer require that operations involve column types not in the typesSet. This allows a user to insert any kind of column types.

For nearly all functions, we only return at one position for the sake of consistency.

We also try to improve stylistic format by settings strings that repeat multiple times to variables to save space.