New option find and ors
There is a new option called Find which allows users to search for patterns that closely match but are not exact matches or are substrings of existing patterns in the key value store. Additionally, OR is now usable as a keyword if the user wants to specify that only a certain set of conditions needs to be met. This is not in conjunctive normal form (ie this would be ORs of ANDs, and might be changed in future pushes).

To implement Find:
printSearchResult now has two new arguments, a boolean indicating whether or not the user specified to "find" similar strings (in the new option called find) and a limit, basically a similarity score between two strings calculated using difflib (see documentation online). Find searches through each pattern as specificied in searches (column values, column types, and keys, if the user requests), and if the pattern supplied per loop has a similarity score above that given ratio (0.8 by default), the result is incorporated, or not filtered out right now.

handleSearches similarly takes in a isFind argument, which indicates if the user specified to find the pattern in the values/types/keys. If the user wants to find the value, the limit is first read in unless none is specified, when the default is set to 0.8.

For each of the SELECTs, SEARCHES, and DELETE options, a separate groupby allows separation of groups in OR clauses, and each clause is processed separately for selecting for items. ANDs would be nested within these clauses separated by ORs.