Process Ands/Ors, Trivials, General Loading
The two functions, processAndandOrs and processAndandOrs2 split the list of items by ands and then filters in each item meeting the criteria. The result is joined together. They are only different functions right now because we need to filter by logic.

We use operator to help us implement trivial projects, and from the module we can use the parse function to select for what we item key we wish to select for.

We can now load in many more json files by searching for the item selected as data, and extracting each column type. Through each type, we add the key and each item for each row in the dataset.