Updating Main and adding New Module InputFile
Main has been updated to take in input files and specify where to find data and columns if possible with a new handleFileInput() function, which also alerts the user if the file does not exist.

First, the user is prompted to type in a file name or hit enter if they want to use the default storage file, though they will be prompted again if they type in a nonexistent file or if they did not specify a file and the storage file (now PistolShrimpDB.bin) does not exist. Then, they are prompted individually to enter key words to where the program will later find columns and data. If they don't type them in, the defaults are columns and data.

In InputFile, there is a class called InputFile with a constructor setting up the filename, key value store to place items in, a boolean indicating if we need to write a new db file, a maximum position indicator described in a prior commit, variables to indicate where to find the columns and data, and a way of indicating file type.

There are now four functions. The main function to use is loadFile, which functions similarly to the old loadFile in main except there is now a check to indicate if the file specified is a bin (with storage data) or .json where data is loaded in from scratch.
setUpDatabase sets up the database by loading in the data from the file, which is then passed to retrieveColumnsData to retrieve or create new columns as well as the extracted data. Next, the key value store is set up by looping through each row in the data and attaching column types to each element.

In retrieveColumnsData, first parse is used to find the location of the column types based on the key the user specified, hopefully somewhere in the metadata and throughout the file.

Next, the data and types are collected using findTheData(), another new function, and we check to make sure the data is not None or is completely empty. If the columns were found, then we grab set up the set of the types used to assign each row's individual data value to a specific data value, making sure to filter out labels that indicate metadata or nonhumanreadable data. Otherwise, we set those type labels to be the set of default values up to the length of the longest row.

findTheData first searches for the data in the list of rows, and then anywhere else in the json file using parse. If the data has not been found, then we search for the largest blob of data at the highest tier. This is a little rudimentary, but we are able to extract some data from this and any further refinement would require the user to specify preferences more, like specifying a location to find the data.
The complication is the types within the  json file. We look for the max sized element for both if the overarching object is in the form of a dictionary (object) and if it is in an array. We set the data to be a single object if the type is neither a dictionary or a list, and if it is in a nested dictionary, we continue to extract the elements until we find viable data. If we have a dictionary remaining after all this as our data, we extract our keys as the types and values as our data and return those.