Created separate class for Insert and more
A few minor changes first. A bug associated with making typesSet a class variable is fixed (it was not properly declared as a class variable when it was in functions).

A separate class to handle most of the actions is now created. This will eventually handle everything the main function used to handle besides the initial loading of the file, and possibly the final writing of the file.

So far, it handles inserts to the temporary key value store, and the processing of save and undo commands.

Insert has three main functions, which are generateRandomKey (which generates a random key and is very simple), generateNewRows which handles creating a new row with the column type associated with its corresponding value.

handleInserts is similar to as described before, only with class variables.