Renaming option enabled
The user can now rename columns. This is available from a new class called renamed. This makes it easy for handleInput to recognized it as a renaming option, and we store the two columns information, which we use to switch from one column to the renamed column. We do so by iterating through each item in the key value store, and setting that specific column's key to be the new one if the old key exists in that row. We keep track of the columns saved so when the user wants to undo the changes, we simply make the modification that we did for each row of the key value store but in reverse.