Implement Search Module and Quick Select Fix
We make sure in Select to prioritize Ands over ors by first splitting the command by ors and then ands.

The value arithmetic is likely not going to get implemented in time, so the code is being cut.

Search is implemented similar to what what was implemented before, but in another module.