Major Fixes (a lot in short span of time)
AndandOrs - processAndandOrs2, which is specifically used by Selects and Deletes, has been fixed. Mainly, the issues were logical, flaws, as it took a little bit of overhaul to change processing from splitting by ands then ors to splitting by ors then ands and iterating through each element in the database just once.

For each element in the dictionary, we include its key in the list of keys that we return for use for writing when, for each specific list of and commands each key meets all the criteria specified by conditions of Select and Delete (see prior commit for those). No extra checks needed for Ors because only one of its subcriteria needs to be met, which we made sure of in the processing of our ands.

In processAndandOrs, the function respeonsible for handling ors/ands in Searches/finds, we apply similar logic in that we specify that for each and clause group all the criteria need to be met, but because Searches/finds have more complicated criteria, that is checked in a new searchList function. Basically, for each pattern and possibly its paired column pattern, each pair of column/value is scanned for a match per row, and if every single pair has a match depending on the exact specification, then we return true, and therefore, include that specific key and row. We finally compile a string listing all the rows as that matches what is currently implemented in the Search class. A lot of old, broken code is commented for now - please ignore that as that is likely getting deleted.

Delete - We make a quick fix to make sure any selecting for key returns a singleton list with a key for consitency and type purposes. We also add a new way of preprocessing input to handle input with ands/ors spearately to create a nested list, handled by processAndandOrs2 in the AndandOrs class. In the case there are no ands/ors, findMatchingKeys handles this.

InputFile - the jsonpathExpr used for searching for our data and colmns now has a customized keyword, but it was not being handled properly (as in the syntax was improper as a .. used to indicate an ancestor was missing).

There are a few stylistic changes, like urging the user to make sure the file is in proper json format, in which case this program will not work and a line length shortening change.

We also make sure we only open files if there is a path to that file.

Insert - We make sure any rows that we add are either can be convereted to a dictionary entry or are already able to be converted. Occasionally string keys will cause errors, specifically malformed node errors.

A function was created (described a little later) which allows us to process all the matches as a string. This addition is specifically to handle entries with whitespace that is necessary.

Main -
Some functions had to be moved to a new module as we can't have modules/classes dependent on one another in a cycle.  A comma was added to facilitate processing, especially for white space. We make sure all variables in concern in the pistolShrimpStore class are self.something variables as there are equivalents in a broader scope and that can cause confusion.

maximumPosition is saved specifically for the instance that a save occurs before the dbFile is created and then we make changes to it again so we don't overwrite or have duplicate entries with maximumPosition.

We now only print the successful rename if a renaming actually occured, and if a column does not exist of the first argument's name, we let the user know.

In the case input is empty, we just want the user to be prompted for a command instead of crashing it.

There were a a few arugment order issues that have been resolved, specifically for handleFileInput's return values. At least one global scope variable with a corresponding variable in pistolShrimpStore class was renamed to further oavoid confusion.

OutputFile - we avoid writing the same row twice when it is being addd by writing it to the end and skipping if it does not have a position.

updateFileWithInserts updates the maximumPosition for the purpose specified above.

Search - AND/Ors are preprocessed to create a list of lists just as specified earlier, and with what was done previously, are all passed to processAndandOrs.

Selects- ands and ors are processed in the same function, and then passed to different functions matching keys and processAndandOrs2 depending on whether they have ands/ors.

Update - is modified to be less restrictive on allowed column types as we are allowed to add columns not existing before.

Finally, we created a new module, SharedFunctions. findMatchingKeys is here (its function described previously). spaceMatches will first have the parser group by the = on the index starting with the one specified by the user, and then the matches are joined and whitespace is trimmed to allow for types/values to have any whitespace on the inside as necessary.

conjMatches is similar, but the grouping is by ors and then ands. The rest of the whitespace preserving and removal occurs on each entry within the andList.