Fixed Searches and Finds and more
Searches and Finds, specifically the parsing of its commands, and the way ands and ors are parsed now work.

Fix 1. Searches/Finds: In processAndandOrs2, which should handle the selection of the keys themselves which fit the command, now interprets the command correctly by only dividing the passed in matches to columns and values when a boolean flag, checkColTypeVal is true. This should be only when the original command contains equal signs and is set in the handleSearches function, which calls this. Otherwise, since the way we set up this function to separate out each clause as a list of criteria that we check off for each key, the criteria is in this case a single element in a list, so we can simply flatten it out.  We pass the arguments from ands and ors and our separated out element(s) to a helper function, searchList, which outputs true if and only if the passed in pattern (and associated column type) match that given in the key value store either in the keys or values. If every call to searchList in handleSearches using every corresponding pattern within the list of values anded together reutrn true, then we include the key.

Fix 2: In handleSearches: in addition to checking if  = is present in the command (described in Fix 1), since handlSEarches uses processAndandOrs, in the case the command does not have ands and ors, list nesting is used in very specific cases to make those commands analagous to having ands and ors.

In addition to those fixes, we started cleanup.
processAndandOrs2 is now selectKeys withAndOrs as that seems more appropriate as it returns keys and handles specifically ands and ors.

In SharedFunctions module, doesColumnTypeValueMatch is a new function that handles the common language between findMatchingKeys (now in sharedFunctions) and selectKeyswithAndsOrs.