Documentation for Shared Function and More
The shared functions module should be fully documented and much of the documentation is implicitly in prior commits. There are a few edits, mainly cleanup with eliminating now obsolete function arguments.

AndsandOrs is no longer a module as it made sense to combine it with SharedFunctions, though the processAndsandOrs function (now searchFilter) is now in Search.py since only Search.py uses this function.

Deadcode (commented code and printSearchResults, now replaced by searchFilter) in Search.py, Delete.py, and SharedFunctions.py has also been removed.

In Delete.py, the key variable is obsolete so it's removed. Also,  in both Delete.py and Selects.py, selectKeyswithAndOrs is in SharedFunctions, so that change was necessary.