Rename and Trivial Selects Patches
Renaming was not being properly saved earlier for columns and values containing a space. Therefore, we added additional code to extract two distinct values in the rename class's constructors (and throwing an error message whenever there can be less or more than two inputs to be extracted).

Additionally, the parser for trivial selects has been modified to allow for numerical characters by using json.dumps to ensure the parser reads each string like a string like it is supposed to.