Some changes and patches
The markdown files are updated with a full status of the project.

Some fixes:
I put a space between ands and ors in Deletes/Selects because strings like Tolland might lead the program to head to the wrong directions in terms of handling and/ors vs commas.

The inputFile might potentially use generateRandomKey someday when setting up the rows/keys in case there are no random values present, so generateRandomKey is moved to sharedFunctions.

The find limit is lowered since 0.8 might be a little too high, making things hard to test.

SharedFunctions now has generateRandomKeys, potentially useful in the future, again.

Also, in the run function, the input is made lowercase to make things easier to parse.